refactor: change false package name jtw => jwt
