+++ backend/src/main/java/ch/sectioninformatique/template/security/SecurityConfiguration.java
@@ -1,6 +1,5 @@
 package ch.sectioninformatique.template.security;
 
-import ch.sectioninformatique.template.jtw.JwtAuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationProvider;
@@ -13,6 +12,8 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import ch.sectioninformatique.template.jwt.JwtAuthenticationFilter;
 
 import java.util.List;
