+++ backend/src/test/java/ch/sectioninformatique/template/auth/AuthControllerTest.java
@@ -1,6 +1,6 @@
 package ch.sectioninformatique.template.auth;
 
-import ch.sectioninformatique.template.jtw.JwtService;
+import ch.sectioninformatique.template.jwt.JwtService;
 import ch.sectioninformatique.template.user.User;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
